split color.hpp
