+++ src/color.h
@@ -36,19 +36,6 @@ public:
 
 };
 
-namespace {
-
-template<typename T>
-T Clamp(const float unclamped, const T max) {
-    if (unclamped < 0)
-        return 0;
-    if (unclamped > 1)
-        return max;
-    return std::round(unclamped * max);
-}
-
-}
-
 class ColorClamped8Bit : public ColorBase<unsigned int> {
 public:
     ColorClamped8Bit() noexcept = default;
@@ -74,15 +61,7 @@ public:
         : ColorBase(mat)
     {}
 
-    ColorClamped8Bit Clamp8Bit() const {
-        const unsigned int MAX_VALUE = 255;
-
-        return ColorClamped8Bit({
-            Clamp<unsigned int>(this->Red(), MAX_VALUE),
-            Clamp<unsigned int>(this->Green(), MAX_VALUE),
-            Clamp<unsigned int>(this->Blue(), MAX_VALUE),
-        });
-    }
+    ColorClamped8Bit Clamp8Bit() const;
 
     friend Color operator+(const Color& a, const Color& b) {
         return Color(a.rgb_ + b.rgb_);
