+++ src/main.cpp
@@ -2,7 +2,7 @@
 
 #include "matrix.hpp"
 #include "ray.hpp"
-#include "canvas.hpp"
+#include "canvas.h"
 
 int main() {
     const float CAMERA_Z = -5;
